Move test_serialization_roundtrip function to utils
